requirements.txt
discord-py-interactions
# git+https://github.com/interactions-py/library@4.1.1-rc1
python-dotenv==0.19.1
requests==2.28.1
interactions-wait-for
dinteractions-Paginator
pymongo
black